requirements.txt
numpy
matplotlib
tqdm
opencv-python
torch
Pillow
torchvision
pyiqa
transformers
open_clip_torch
pyyaml
bpy
mathutils